Move HNSW metrics to new submodule
