+++ iris-mpc-cpu/src/hnsw/mod.rs
@@ -1,3 +1,4 @@
 pub mod searcher;
+pub mod metrics;
 
 pub use searcher::HnswSearcher;
