+++ iris-mpc-cpu/src/hnsw/searcher.rs
@@ -4,7 +4,7 @@
 //*
 //* https://github.com/Inversed-Tech/hawk-pack/
 
-use ::tracing::info;
+use tracing::{info, instrument};
 pub use hawk_pack::data_structures::queue::{
     FurthestQueue, FurthestQueueV, NearestQueue, NearestQueueV,
 };
@@ -13,6 +13,8 @@ use rand::RngCore;
 use rand_distr::{Distribution, Geometric};
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
+
+use super::metrics;
 
 // specify construction and search parameters by layer up to this value minus 1
 // any higher layers will use the last set of parameters
@@ -217,6 +219,7 @@ impl HnswSearcher {
     /// given layer using depth-first graph traversal,  Terminates when `W`
     /// contains vectors which are the nearest to `q` among all traversed
     /// vertices and their neighbors.
+    #[instrument(skip(self, vector_store, graph_store))]
     #[allow(non_snake_case)]
     async fn search_layer<V: VectorStore, G: GraphStore<V>>(
         &self,
@@ -227,7 +230,7 @@ impl HnswSearcher {
         ef: usize,
         lc: usize,
     ) {
-        info!(event_type = tracing::LAYER_SEARCH_EVENT);
+        info!(event_type = metrics::LAYER_SEARCH_EVENT);
 
         // v: The set of already visited vectors.
         let mut v = HashSet::<V::VectorRef>::from_iter(W.iter().map(|(e, _eq)| e.clone()));
@@ -248,7 +251,7 @@ impl HnswSearcher {
             }
 
             // Open the node c and explore its neighbors.
-            info!(event_type = tracing::OPEN_NODE_EVENT);
+            info!(event_type = metrics::OPEN_NODE_EVENT);
 
             // Visit all neighbors of c.
             let c_links = graph_store.get_links(&c, lc).await;
@@ -441,71 +444,6 @@ impl HnswSearcher {
     }
 }
 
-pub mod tracing {
-    use std::sync::{
-        atomic::{AtomicUsize, Ordering},
-        Arc,
-    };
-    use tracing::{Event, Subscriber};
-    use tracing_subscriber::{layer::Context, Layer};
-
-    pub const LAYER_SEARCH_EVENT: u64 = 0;
-    pub const OPEN_NODE_EVENT: u64 = 1;
-    pub const EVAL_DIST_EVENT: u64 = 2;
-    pub const COMPARE_DIST_EVENT: u64 = 3;
-
-    const NUM_EVENT_TYPES: usize = 4;
-
-    #[derive(Default)]
-    pub struct EventCounter {
-        pub counters: [AtomicUsize; NUM_EVENT_TYPES],
-    }
-
-    pub struct HnswEventCounterLayer {
-        pub counters: Arc<EventCounter>,
-    }
-
-    impl<S: Subscriber> Layer<S> for HnswEventCounterLayer {
-        fn on_event(&self, event: &Event<'_>, _ctx: Context<'_, S>) {
-            let mut visitor = EventVisitor::default();
-            event.record(&mut visitor);
-
-            if let Some(event_type) = visitor.event {
-                if let Some(counter) = self.counters.counters.get(event_type) {
-                    let increment_amount = visitor.amount.unwrap_or(1);
-                    counter.fetch_add(increment_amount, Ordering::Relaxed);
-                } else {
-                    panic!("Invalid event type specified: {:?}", event_type);
-                }
-            }
-        }
-    }
-
-    #[derive(Default)]
-    struct EventVisitor {
-        // which event was encountered
-        event: Option<usize>,
-
-        // how much to increment the associated counter
-        amount: Option<usize>,
-    }
-
-    impl tracing::field::Visit for EventVisitor {
-        fn record_u64(&mut self, field: &tracing::field::Field, value: u64) {
-            match field.name() {
-                "event_type" => {
-                    self.event = Some(value as usize);
-                }
-                "increment_amount" => {
-                    self.amount = Some(value as usize);
-                }
-                _ => {}
-            }
-        }
-
-        fn record_debug(&mut self, _field: &tracing::field::Field, _value: &dyn std::fmt::Debug) {}
-    }
-}
 
 #[cfg(test)]
 mod tests {
